Rydder i pakkestruktur i tiltaksmodulen
