+++ tiltak/src/main/kotlin/no/nav/amt/tiltak/deltaker/commands/UpsertNavAnsattCommand.kt
@@ -1,0 +1,8 @@
+package no.nav.amt.tiltak.deltaker.commands
+
+data class UpsertNavAnsattCommand(
+	val personligIdent: String,
+	val navn: String,
+	val telefonnummer: String?,
+	val epost: String?
+)
