+++ tiltak/src/main/kotlin/no/nav/amt/tiltak/deltaker/dbo/BrukerDbo.kt
@@ -1,0 +1,17 @@
+package no.nav.amt.tiltak.deltaker.dbo
+
+import java.time.LocalDateTime
+import java.util.*
+
+data class BrukerDbo(
+	val id: UUID,
+	val fodselsnummer: String,
+	val fornavn: String,
+	val mellomnavn: String?,
+	val etternavn: String,
+	val telefonnummer: String?,
+	val epost: String?,
+	val ansvarligVeilederId: UUID?,
+	val createdAt: LocalDateTime,
+	val modifiedAt: LocalDateTime
+)
