--- tiltak/src/main/kotlin/no/nav/amt/tiltak/tiltak/controllers/dto/TiltakDto.kt
@@ -1,13 +1,0 @@
-package no.nav.amt.tiltak.tiltak.controllers.dto
-
-import no.nav.amt.tiltak.core.domain.tiltak.Tiltak
-
-data class TiltakDto(
-	val tiltakskode: String,
-	val tiltaksnavn: String,
-)
-
-fun Tiltak.toDto() = TiltakDto(
-	tiltakskode = this.kode,
-	tiltaksnavn = this.navn
-)
