+++ tiltak/src/main/kotlin/no/nav/amt/tiltak/tiltak/dbo/TiltakDbo.kt
@@ -1,8 +1,8 @@
 package no.nav.amt.tiltak.tiltak.dbo
 
 import no.nav.amt.tiltak.core.domain.tiltak.Tiltak
-import no.nav.amt.tiltak.tiltak.utils.UpdateCheck
-import no.nav.amt.tiltak.tiltak.utils.UpdateStatus
+import no.nav.amt.tiltak.utils.UpdateCheck
+import no.nav.amt.tiltak.utils.UpdateStatus
 import java.time.LocalDateTime
 import java.util.*
