--- tiltak/src/main/kotlin/no/nav/amt/tiltak/tiltak/deltaker/commands/UpsertNavAnsattCommand.kt
@@ -1,8 +1,0 @@
-package no.nav.amt.tiltak.tiltak.deltaker.commands
-
-data class UpsertNavAnsattCommand(
-	val personligIdent: String,
-	val navn: String,
-	val telefonnummer: String?,
-	val epost: String?
-)
