--- tiltak/src/main/kotlin/no/nav/amt/tiltak/tiltak/deltaker/dbo/NavAnsattDbo.kt
@@ -1,11 +1,0 @@
-package no.nav.amt.tiltak.tiltak.deltaker.dbo
-
-import java.util.*
-
-data class NavAnsattDbo(
-	val id: UUID,
-	val personligIdent: String,
-	val navn: String,
-	val telefonnummer: String?,
-	val epost: String?
-)
