+++ tiltak/src/main/kotlin/no/nav/amt/tiltak/tiltak/dto/TiltakDeltakerDto.kt
@@ -1,0 +1,26 @@
+package no.nav.amt.tiltak.tiltak.dto
+
+import no.nav.amt.tiltak.core.domain.tiltak.Deltaker
+import java.time.LocalDate
+import java.util.*
+
+data class TiltakDeltakerDto(
+	val id: UUID,
+	val fornavn: String,
+	val mellomnavn: String? = null,
+	val etternavn: String,
+	val fodselsnummer: String,
+	val startdato: LocalDate?,
+	val sluttdato: LocalDate?,
+	val status: Deltaker.Status?
+)
+
+fun Deltaker.toDto() = TiltakDeltakerDto(
+	id = id,
+	fornavn = fornavn,
+	etternavn = etternavn,
+	fodselsnummer = fodselsnummer,
+	startdato = startdato,
+	sluttdato = sluttdato,
+	status = status
+)
