+++ tiltak/src/main/kotlin/no/nav/amt/tiltak/tiltak/dto/TiltakDto.kt
@@ -1,0 +1,13 @@
+package no.nav.amt.tiltak.tiltak.dto
+
+import no.nav.amt.tiltak.core.domain.tiltak.Tiltak
+
+data class TiltakDto(
+	val tiltakskode: String,
+	val tiltaksnavn: String,
+)
+
+fun Tiltak.toDto() = TiltakDto(
+	tiltakskode = this.kode,
+	tiltaksnavn = this.navn
+)
