+++ tiltak/src/main/kotlin/no/nav/amt/tiltak/tiltak/services/DeltakerServiceImpl.kt
@@ -3,13 +3,13 @@ package no.nav.amt.tiltak.tiltak.services
 import no.nav.amt.tiltak.core.domain.tiltak.Deltaker
 import no.nav.amt.tiltak.core.port.DeltakerService
 import no.nav.amt.tiltak.core.port.PersonService
-import no.nav.amt.tiltak.tiltak.deltaker.commands.UpsertNavAnsattCommand
-import no.nav.amt.tiltak.tiltak.deltaker.dbo.BrukerDbo
-import no.nav.amt.tiltak.tiltak.deltaker.dbo.NavAnsattDbo
-import no.nav.amt.tiltak.tiltak.deltaker.repositories.BrukerRepository
-import no.nav.amt.tiltak.tiltak.deltaker.repositories.DeltakerRepository
-import no.nav.amt.tiltak.tiltak.deltaker.repositories.NavAnsattRepository
-import no.nav.amt.tiltak.tiltak.utils.UpdateStatus
+import no.nav.amt.tiltak.deltaker.commands.UpsertNavAnsattCommand
+import no.nav.amt.tiltak.deltaker.dbo.BrukerDbo
+import no.nav.amt.tiltak.deltaker.dbo.NavAnsattDbo
+import no.nav.amt.tiltak.deltaker.repositories.BrukerRepository
+import no.nav.amt.tiltak.deltaker.repositories.DeltakerRepository
+import no.nav.amt.tiltak.deltaker.repositories.NavAnsattRepository
+import no.nav.amt.tiltak.utils.UpdateStatus
 import org.springframework.stereotype.Service
 import java.time.LocalDate
 import java.util.*
