--- tiltak/src/main/kotlin/no/nav/amt/tiltak/tiltak/utils/UpdateCheck.kt
@@ -1,11 +1,0 @@
-package no.nav.amt.tiltak.tiltak.utils
-
-enum class UpdateStatus {
-	UPDATED,
-	NO_CHANGE
-}
-
-data class UpdateCheck<T>(
-	val status: UpdateStatus,
-	val updatedObject: T? = null
-)
