+++ tiltak/src/main/kotlin/no/nav/amt/tiltak/utils/UpdateCheck.kt
@@ -1,0 +1,11 @@
+package no.nav.amt.tiltak.utils
+
+enum class UpdateStatus {
+	UPDATED,
+	NO_CHANGE
+}
+
+data class UpdateCheck<T>(
+	val status: UpdateStatus,
+	val updatedObject: T? = null
+)
