+++ tiltak/src/test/kotlin/no/nav/amt/tiltak/TestApplication.kt
@@ -1,0 +1,8 @@
+package no.nav.amt.tiltak
+
+import no.nav.security.token.support.spring.api.EnableJwtTokenValidation
+import org.springframework.boot.autoconfigure.SpringBootApplication
+
+@EnableJwtTokenValidation
+@SpringBootApplication
+open class TestApplication
