--- tiltak/src/test/kotlin/no/nav/amt/tiltak/tiltak/TestApplication.kt
@@ -1,8 +1,0 @@
-package no.nav.amt.tiltak.tiltak
-
-import no.nav.security.token.support.spring.api.EnableJwtTokenValidation
-import org.springframework.boot.autoconfigure.SpringBootApplication
-
-@EnableJwtTokenValidation
-@SpringBootApplication
-open class TestApplication
