+++ tiltak/src/test/kotlin/no/nav/amt/tiltak/tiltak/controllers/TiltakInstansControllerIntegrationTest.kt
@@ -8,10 +8,10 @@ import no.nav.amt.tiltak.core.port.TiltakInstansService
 import no.nav.amt.tiltak.test.database.DatabaseTestUtils
 import no.nav.amt.tiltak.test.database.SingletonPostgresContainer
 import no.nav.amt.tiltak.tiltak.dbo.TiltakInstansDbo
-import no.nav.amt.tiltak.tiltak.deltaker.dbo.DeltakerDbo
-import no.nav.amt.tiltak.tiltak.deltaker.repositories.BrukerRepository
-import no.nav.amt.tiltak.tiltak.deltaker.repositories.DeltakerRepository
-import no.nav.amt.tiltak.tiltak.deltaker.repositories.NavAnsattRepository
+import no.nav.amt.tiltak.deltaker.dbo.DeltakerDbo
+import no.nav.amt.tiltak.deltaker.repositories.BrukerRepository
+import no.nav.amt.tiltak.deltaker.repositories.DeltakerRepository
+import no.nav.amt.tiltak.deltaker.repositories.NavAnsattRepository
 import no.nav.amt.tiltak.tiltak.repositories.TiltakInstansRepository
 import no.nav.amt.tiltak.tiltak.repositories.TiltakRepository
 import no.nav.amt.tiltak.tiltak.services.DeltakerServiceImpl
@@ -40,7 +40,7 @@ class TiltakInstansControllerIntegrationTest {
 	private lateinit var tiltakInstansRepository: TiltakInstansRepository
 	private lateinit var tiltakInstansService: TiltakInstansService
 	private lateinit var deltakerService: DeltakerService
-	private lateinit var controller: TiltakInstansController
+	private lateinit var controller: no.nav.amt.tiltak.tiltak.controllers.TiltakInstansController
 
 	@BeforeEach
 	fun before() {
@@ -57,7 +57,7 @@ class TiltakInstansControllerIntegrationTest {
 			mock(PersonService::class.java)
 		);
 		tiltakInstansService = TiltakInstansServiceImpl(tiltakInstansRepository, TiltakServiceImpl(tiltakRepository))
-		controller = TiltakInstansController(tiltakInstansService, deltakerService)
+		controller = no.nav.amt.tiltak.tiltak.controllers.TiltakInstansController(tiltakInstansService, deltakerService)
 
 		DatabaseTestUtils.cleanDatabase(dataSource)
