+++ tiltaksleverandor/src/main/kotlin/no/nav/amt/tiltak/tiltaksleverandor/dbo/TiltaksleverandorDbo.kt
@@ -1,8 +1,8 @@
 package no.nav.amt.tiltak.tiltaksleverandor.dbo
 
 import no.nav.amt.tiltak.core.domain.tiltaksleverandor.Tiltaksleverandor
-import no.nav.amt.tiltak.tiltak.utils.UpdateCheck
-import no.nav.amt.tiltak.tiltak.utils.UpdateStatus
+import no.nav.amt.tiltak.utils.UpdateCheck
+import no.nav.amt.tiltak.utils.UpdateStatus
 import java.time.LocalDateTime
 import java.util.*
